requirements.txt
discord.py==1.7.3
python-dotenv==0.18.0
pandas==1.2.4
nest-asyncio==1.5.1